lib/models/TIme_series/requirements.txt
numpy
pandas
shapely
python-dotenv
sentinelhub
scikit-learn
matplotlib
neuralforecast
openmeteo-requests
requests-cache
retry-requests
openmeteo-sdk
ray[tune]
